+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 * `ComponentOrNode` moved to test module. Renamed to `ReactOrDomNode`.
 * Changed overloaded `classSet` methods into `classSet{,1}{,M}`.
+* Added `simulateKeyDownUp` and `simulateKeyDownPressUp` to `KeyboardEventData` in the test module.
 
 # 0.6.0 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.4...v0.6.0))